lib/ushell: fix context switch from the main thred
Previously if ushell was context-switched from the main thread that was
sleeping, the thread information was still in the sleeping threads list.
This causes scheduling problems if ushell yield() and called scheduler
functions (e.g., by using a wait queue for waiting for interrupts of
virtio devices). Fix this by changing the thread state at the beginning
and the end of the ushell entry function.