Execute ushell in a seperate thread
This patch enables threading for ushell. After the initialization of
Unikraft (platform, drivers, libs, etc.) a new htread is sapwned for
ushell. The thread simply communicates with the ukconsoledev library,
where it receives all the data.

If there are no data, then the thread waits in a semaphore. As soon as,
the virtio console driver gets new data, it stores them and increases
the semaphore, waking up the ushell thread.

In the same time the application continues its execution. However, after
the reception of a new packet, the application needs to get stopped, in
order ushell thread to get scheduled. Since, the current scheduler is
cooperative, the application thread needs to yield and can not get
interrupted.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>